pyproject.toml
[project]
name = "art-director-mcp"
version = "0.1.0"
description = "MCP Server that orchestrates a high-fidelity image generation pipeline with AI-powered planning, execution, and quality auditing."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Art Director" }]

dependencies = [
    "mcp[cli]>=1.26.0",
    "openai>=1.40.0",
    "huggingface-hub>=0.25.0",
    "httpx>=0.27.0",
    "Pillow>=10.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
    "freezegun>=1.3.0",
    "ruff>=0.6.0",
]

[project.scripts]
art-director = "art_director.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/art_director"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
